Moved JSONParser to Parsing module
