+++ src/CMakeLists.txt
@@ -3,6 +3,7 @@ add_subdirectory(Events)
 add_subdirectory(Game)
 add_subdirectory(IO)
 add_subdirectory(Logging)
+add_subdirectory(Parsing)
 add_subdirectory(Utils)
 add_executable(espadaNegra main.cpp)
 
@@ -12,5 +13,6 @@ PRIVATE
   Game
   IO 
   Logging
+  Parsing
   Utils
 )
