+++ src/Parsing/CMakeLists.txt
@@ -1,0 +1,14 @@
+add_library(Parsing
+    JSONParser.cpp
+    JSONParser.hpp
+)
+
+find_package(Boost REQUIRED)
+include_directories(${Boost_INCLUDE_DIR})
+target_link_libraries(Parsing 
+PRIVATE 
+  ${Boost_LIBRARIES} 
+  Logging 
+  Utils 
+  Game
+)
