+++ src/Parsing/JSONParser.cpp
@@ -9,32 +9,32 @@
 #include <memory>
 #include <string>
 
-#include "AttackRequest.hpp"
-#include "AttackResult.hpp"
-#include "Draw.hpp"
-#include "Earn.hpp"
-#include "Error.hpp"
+#include "Events/AttackRequest.hpp"
+#include "Events/AttackResult.hpp"
+#include "Events/Draw.hpp"
+#include "Events/Earn.hpp"
+#include "Events/Error.hpp"
+#include "Events/GetCard.hpp"
+#include "Events/InputEvent.hpp"
+#include "Events/IsAuctioneer.hpp"
+#include "Events/JoinMatchRequest.hpp"
+#include "Events/JoinMatchResult.hpp"
+#include "Events/Leave.hpp"
+#include "Events/MatchInfo.hpp"
+#include "Events/OfferRequest.hpp"
+#include "Events/OfferResult.hpp"
+#include "Events/OutputEvent.hpp"
+#include "Events/PassRequest.hpp"
+#include "Events/PassResult.hpp"
+#include "Events/Pay.hpp"
+#include "Events/SetGold.hpp"
+#include "Events/Winner.hpp"
 #include "Game/Card.hpp"
 #include "Game/Deck.hpp"
 #include "Game/MatchConfig.hpp"
 #include "Game/Player.hpp"
-#include "GetCard.hpp"
-#include "InputEvent.hpp"
-#include "IsAuctioneer.hpp"
-#include "JoinMatchRequest.hpp"
-#include "JoinMatchResult.hpp"
-#include "Leave.hpp"
 #include "Logging/Logger.hpp"
-#include "MatchInfo.hpp"
-#include "OfferRequest.hpp"
-#include "OfferResult.hpp"
-#include "OutputEvent.hpp"
-#include "PassRequest.hpp"
-#include "PassResult.hpp"
-#include "Pay.hpp"
-#include "SetGold.hpp"
 #include "Utils/Time.hpp"
-#include "Winner.hpp"
 
 using IT = InputEvent::Type;
 using OT = OutputEvent::Type;
